lab4/requirements.txt
grpcio==1.47.5
grpcio-tools==1.47.5
protobuf==3.20.3
psutil==5.9.6
etcd3==0.12.0
kafka-python==2.0.2
